+++ native/packages/react-native-bpk-component-switch/src/BpkSwitch-test.common.js
@@ -1,0 +1,73 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2017 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import renderer from 'react-test-renderer';
+import BpkThemeProvider from 'react-native-bpk-theming';
+
+import BpkSwitch, { propTypes } from './BpkSwitch';
+
+const commonTests = () => {
+  describe('BpkSwitch', () => {
+    // Currently necessary because of https://github.com/facebook/react-native/issues/16247
+    jest.spyOn(console, 'error').mockImplementation(() => jest.fn());
+    it('should render correctly', () => {
+      const tree = renderer.create(
+        <BpkSwitch />,
+      ).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+
+    it('should support the "value" prop', () => {
+      const tree = renderer.create(
+        <BpkSwitch value />,
+      ).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+
+    it('should support theming', () => {
+      const theme = {
+        switchPrimaryColor: 'red',
+      };
+      const tree = renderer.create(
+        <BpkThemeProvider theme={theme}>
+          <BpkSwitch />
+        </BpkThemeProvider>,
+      ).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+
+    it('should reject theme property when required theme attributes are omitted', () => { // eslint-disable-line max-len
+      expect(propTypes.theme({
+        type: 'primary',
+        theme: {},
+      }, 'theme', 'BpkSwitch').toString()).toEqual('Error: Invalid prop `theme` supplied to `BpkSwitch`. For themed switches, `theme` prop must include `switchPrimaryColor`'); // eslint-disable-line max-len
+    });
+
+    it('should disable theming if the required attribute is omitted', () => {
+      const theme = {};
+      const tree = renderer.create(
+        <BpkThemeProvider theme={theme}>
+          <BpkSwitch />
+        </BpkThemeProvider>,
+      ).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+  });
+};
+export default commonTests;